% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bboread.R
\name{bboread}
\alias{bboread}
\title{Read in BBO data sets from CME Historical Datamine}
\usage{
bboread(x)
}
\arguments{
\item{x}{is a .txt file from the CME Group's historical datamine.}
}
\value{
A datatable.
}
\description{
This function is simply a wrapper for the read_fwf function in the readr package. Inside the function, start and end positions
of the raw fixed width file are specified as documented at
http://www.cmegroup.com/confluence/display/EPICSANDBOX/Best+Bid+Best+Offer+ASCII+Layout
For performance, only the bare minimum columns are read in.
}
\examples{
DATA <- bboread('data-raw/XCBT_C_FUT_110110.TXT')
head(DATA)
}

