\name{BBOTools-package}
\alias{BBOTools-package}
\alias{BBOTools}
\docType{package}
\title{
\packageTitle{BBOTools}
}
\description{
\packageDescription{BBOTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BBOTools}
\packageIndices{BBOTools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BBOTools}

Maintainer: \packageMaintainer{BBOTools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
